[FIX] pos_restaurant: ensure pos order is removed from local storage
Before this commit, when order had no orderlines and guest count was set to 0, removeOrder was called which deleted this order from local storage. Because guest count was set to 0, method _onReactiveOrderUpdated was also triggered and deleted order was again saved to local storage. With this commit we lock the order so that method save_to_db does not save.